autenticacion/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.7.0